requirements/requirements.txt
PyYAML>=6.0
numpy>=1.26
tqdm>=4.66
torch>=2.2
